// web/src/api/models/shared-collection.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SharedCollectionConfig } from './shared-collection-config';

/**
 * Shared Collection information for marketplace users
 * @export
 * @interface SharedCollection
 */
export interface SharedCollection {
    /**
     * Collection ID
     * @type {string}
     * @memberof SharedCollection
     */
    'id': string;
    /**
     * Collection title
     * @type {string}
     * @memberof SharedCollection
     */
    'title': string;
    /**
     * Collection description
     * @type {string}
     * @memberof SharedCollection
     */
    'description'?: string;
    /**
     * Original owner user ID
     * @type {string}
     * @memberof SharedCollection
     */
    'owner_user_id': string;
    /**
     * Original owner username
     * @type {string}
     * @memberof SharedCollection
     */
    'owner_username'?: string;
    /**
     * Subscription record ID (has value if subscribed, null if not subscribed)
     * @type {string}
     * @memberof SharedCollection
     */
    'subscription_id'?: string;
    /**
     * Subscription time (only has value when subscribed)
     * @type {string}
     * @memberof SharedCollection
     */
    'gmt_subscribed'?: string;
    /**
     * Total number of subscriptions
     * @type {number}
     * @memberof SharedCollection
     */
    'subscription_count': number;
    /**
     * Collection configuration settings
     * @type {SharedCollectionConfig}
     * @memberof SharedCollection
     */
    'config': SharedCollectionConfig;
}